# web/openapi.json
{
  "components": {
    "schemas": {
      "AgeCurveOut": {
        "description": "One stat's delta-method aging curve.",
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "n_pairs": {
            "title": "N Pairs",
            "type": "integer"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/AgePointOut"
            },
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "metric",
          "label",
          "n_pairs",
          "points"
        ],
        "title": "AgeCurveOut",
        "type": "object"
      },
      "AgeCurvesResponse": {
        "description": "Aging curves from past FPL seasons (PRD \u00a78.10 step 4, US-18).",
        "properties": {
          "caveat": {
            "title": "Caveat",
            "type": "string"
          },
          "curves": {
            "items": {
              "$ref": "#/components/schemas/AgeCurveOut"
            },
            "title": "Curves",
            "type": "array"
          },
          "history_as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "History As Of"
          },
          "min_minutes": {
            "title": "Min Minutes",
            "type": "number"
          },
          "min_pairs": {
            "title": "Min Pairs",
            "type": "integer"
          },
          "pair_count": {
            "title": "Pair Count",
            "type": "integer"
          },
          "seasons": {
            "items": {
              "type": "string"
            },
            "title": "Seasons",
            "type": "array"
          }
        },
        "required": [
          "seasons",
          "min_minutes",
          "min_pairs",
          "pair_count",
          "history_as_of",
          "caveat",
          "curves"
        ],
        "title": "AgeCurvesResponse",
        "type": "object"
      },
...
      "BacktestResponse": {
        "description": "Last season's diagnosis vs this season's arrivals (US-16, exploratory).",
        "properties": {
          "arrivals_as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrivals As Of"
          },
          "as_of_season": {
            "description": "Season the needs were diagnosed at the end of.",
            "title": "As Of Season",
            "type": "string"
          },
          "baseline_precision": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Same for the n most-signed groups at the other clubs.",
            "title": "Baseline Precision"
          },
          "benchmark": {
            "title": "Benchmark",
            "type": "string"
          },
          "caveat": {
            "title": "Caveat",
            "type": "string"
          },
          "clubs": {
            "items": {
              "$ref": "#/components/schemas/ClubBacktestOut"
            },
            "title": "Clubs",
            "type": "array"
          },
          "evaluated": {
            "title": "Evaluated",
            "type": "integer"
          },
          "history_as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "History As Of"
          },
          "hit_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Share of clubs with at least one hit.",
            "title": "Hit Rate"
          },
          "precision": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Mean precision@n over evaluated clubs.",
            "title": "Precision"
          },
          "random_precision": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Expected precision of random groups.",
            "title": "Random Precision"
          },
          "signing_season": {
            "description": "Season whose arrivals are the outcome.",
            "title": "Signing Season",
            "type": "string"
          },
          "skipped": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Skipped",
            "type": "object"
          },
          "top_n": {
            "title": "Top N",
            "type": "integer"
          }
        },
        "required": [
          "as_of_season",
          "signing_season",
          "benchmark",
          "top_n",
          "precision",
          "baseline_precision",
          "random_precision",
          "hit_rate",
          "evaluated",
          "skipped",
          "clubs",
          "history_as_of",
          "arrivals_as_of",
          "caveat"
        ],
        "title": "BacktestResponse",
        "type": "object"
      },
...
      "ClubBacktestOut": {
        "description": "Predicted needs vs actual arrivals for one club.",
        "properties": {
          "arrivals": {
            "items": {
              "$ref": "#/components/schemas/ArrivalOut"
            },
            "title": "Arrivals",
            "type": "array"
          },
          "baseline_groups": {
            "items": {
              "type": "string"
            },
            "title": "Baseline Groups",
            "type": "array"
          },
          "baseline_precision": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Precision"
          },
          "hits": {
            "items": {
              "type": "string"
            },
            "title": "Hits",
            "type": "array"
          },
          "precision": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precision"
          },
          "predicted": {
            "items": {
              "$ref": "#/components/schemas/PredictedNeedOut"
            },
            "title": "Predicted",
            "type": "array"
          },
          "random_precision": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Random Precision"
          },
          "team_id": {
            "title": "Team Id",
            "type": "integer"
          },
          "team_name": {
            "title": "Team Name",
            "type": "string"
          }
        },
        "required": [
          "team_id",
          "team_name",
          "predicted",
          "arrivals",
          "hits",
          "precision",
          "baseline_groups",
          "baseline_precision",
          "random_precision"
        ],
        "title": "ClubBacktestOut",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Machine-readable error details.",
        "properties": {
          "code": {
            "description": "Stable error code, e.g. 'not_found'.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "The single error schema returned by every endpoint.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FreshnessResponse": {
        "description": "Per-source freshness, mapping coverage and validation status (US-08).",
        "properties": {
          "coverage": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Share of FPL minutes mapped per source.",
            "title": "Coverage",
            "type": "object"
          },
          "fpl_schema": {
            "title": "Fpl Schema",
            "type": "string"
          },
          "review_count": {
            "description": "Records entity resolution could not map.",
            "title": "Review Count",
            "type": "integer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceFreshnessOut"
            },
            "title": "Sources",
            "type": "array"
          },
          "validation_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Summary"
          },
          "warehouse_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warehouse Version"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "warehouse_version",
          "validation_summary",
          "sources",
          "coverage",
          "review_count",
          "fpl_schema",
          "warnings"
        ],
        "title": "FreshnessResponse",
        "type": "object"
      },
...
      "KpiDefinitionOut": {
        "description": "One player KPI definition from ``config/kpis.yaml``.",
        "properties": {
          "higher_is_better": {
            "title": "Higher Is Better",
            "type": "boolean"
          },
          "is_proxy": {
            "title": "Is Proxy",
            "type": "boolean"
          },
          "kpi": {
            "title": "Kpi",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "possession_adjusted": {
            "title": "Possession Adjusted",
            "type": "boolean"
          },
          "proxy_for": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy For"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "kpi",
          "label",
          "source",
          "higher_is_better",
          "is_proxy",
          "proxy_for",
          "possession_adjusted"
        ],
        "title": "KpiDefinitionOut",
        "type": "object"
      },
...
      "MethodologyResponse": {
        "description": "Definitions, weights, thresholds and limitations (US-08, PRD \u00a78).",
        "properties": {
          "fit_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Fit Weights",
            "type": "object"
          },
          "group_caveats": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Groups whose free metrics are limited.",
            "title": "Group Caveats",
            "type": "object"
          },
          "kpis": {
            "items": {
              "$ref": "#/components/schemas/KpiDefinitionOut"
            },
            "title": "Kpis",
            "type": "array"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelRunOut"
            },
            "title": "Models",
            "type": "array"
          },
          "parameters": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                }
              ]
            },
            "description": "Methodology and diagnosis thresholds from config/settings.yaml.",
            "title": "Parameters",
            "type": "object"
          },
          "peak_age": {
            "additionalProperties": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "integer"
                },
                {
                  "type": "integer"
                }
              ],
              "type": "array"
            },
            "title": "Peak Age",
            "type": "object"
          },
          "percentile_bands": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Percentile Bands",
            "type": "object"
          },
          "position_groups": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "number"
              },
              "type": "object"
            },
            "description": "Group -> KPI -> weight.",
            "title": "Position Groups",
            "type": "object"
          },
          "team_kpis": {
            "items": {
              "$ref": "#/components/schemas/TeamKpiDefinitionOut"
            },
            "title": "Team Kpis",
            "type": "array"
          },
          "upgrade_gate_min_delta": {
            "title": "Upgrade Gate Min Delta",
            "type": "number"
          },
          "value_model_caveat": {
            "title": "Value Model Caveat",
            "type": "string"
          }
        },
        "required": [
          "kpis",
          "position_groups",
          "group_caveats",
          "team_kpis",
          "fit_weights",
          "upgrade_gate_min_delta",
          "peak_age",
          "parameters",
          "percentile_bands",
          "value_model_caveat",
          "models",
          "limitations"
        ],
        "title": "MethodologyResponse",
        "type": "object"
      },
      "ModelRunOut": {
        "description": "When an ML output stored in the warehouse was produced.",
        "properties": {
          "git_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Sha"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rows": {
            "title": "Rows",
            "type": "integer"
          },
          "trained_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trained At"
          }
        },
        "required": [
          "name",
          "trained_at",
          "git_sha",
          "rows"
        ],
        "title": "ModelRunOut",
        "type": "object"
      },
...
      "SourceFreshnessOut": {
        "description": "Newest snapshot of one source.",
        "properties": {
          "age_hours": {
            "title": "Age Hours",
            "type": "number"
          },
          "last_fetched": {
            "format": "date-time",
            "title": "Last Fetched",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          }
        },
        "required": [
          "source",
          "last_fetched",
          "age_hours",
          "stale"
        ],
        "title": "SourceFreshnessOut",
        "type": "object"
      },
      "TeamKpiDefinitionOut": {
        "description": "One team KPI and the position groups a shortfall maps to.",
        "properties": {
          "higher_is_better": {
            "title": "Higher Is Better",
            "type": "boolean"
          },
          "kpi": {
            "title": "Kpi",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "responsible_groups": {
            "items": {
              "type": "string"
            },
            "title": "Responsible Groups",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "kpi",
          "label",
          "source",
          "higher_is_better",
          "responsible_groups"
        ],
        "title": "TeamKpiDefinitionOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Read-only API over the local warehouse. No number without a receipt.",
    "title": "PL Scout Engine API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/meta/age-curves": {
      "get": {
        "description": "Delta-method aging curves per stat from past FPL seasons (US-18).",
        "operationId": "age_curve_list_meta_age_curves_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgeCurvesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown club, player or need."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request parameters."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The warehouse has not been built yet."
          }
        },
        "summary": "Age Curve List",
        "tags": [
          "meta"
        ]
      }
    },
    "/meta/backtest": {
      "get": {
        "description": "Did last season's top needs match the position groups clubs then signed? (US-16).",
        "operationId": "backtest_meta_backtest_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown club, player or need."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request parameters."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The warehouse has not been built yet."
          }
        },
        "summary": "Backtest",
        "tags": [
          "meta"
        ]
      }
    },
    "/meta/freshness": {
      "get": {
        "description": "Per-source last fetch, mapping coverage, staleness flags and validation status.",
        "operationId": "freshness_meta_freshness_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FreshnessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown club, player or need."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request parameters."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The warehouse has not been built yet."
          }
        },
        "summary": "Freshness",
        "tags": [
          "meta"
        ]
      }
    },
    "/meta/methodology": {
      "get": {
        "description": "KPI definitions, weights, proxies, thresholds, model runs and known limitations.",
        "operationId": "methodology_meta_methodology_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MethodologyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown club, player or need."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request parameters."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The warehouse has not been built yet."
          }
        },
        "summary": "Methodology",
        "tags": [
          "meta"
        ]
      }
    },
...
  }
}